[docs] Improve /customization/typography/
