+++ docs/src/pages/customization/breakpoints/breakpoints.md
@@ -261,3 +261,7 @@ function MyComponent(props) {
 
 export default withWidth()(MyComponent);
 ```
+
+## The default values
+
+You can explore all the default values of the breakpoints using our [explorer](/customization/default-theme/?expend-path=$.breakpoints) or by opening the dev tools console on this page (`window.theme.breakpoints`).
