+++ docs/src/pages/customization/spacing/spacing.md
@@ -5,13 +5,12 @@
 Material-UI uses [a recommended 8px scaling factor by default](https://material.io/design/layout/understanding-layout.html).
 
 ```js
-const styles = theme => ({
-  root: {
-    // JSS uses px as the default units for this CSS property.
-    padding: theme.spacing(2), // = 8 * 2
-  },
-});
+const theme = createMuiTheme();
+
+theme.spacing(2) // = 8 * 2
 ```
+
+## Custom spacing
 
 You can change the spacing transformation by providing:
 
@@ -48,8 +47,14 @@ theme.spacing(2); // = 8
 ## Multiple arity
 
 The `theme.spacing()` helper accepts up to 4 arguments.
-You can use the arguments to reduce the boilerplate:
-```diff
--  padding: `${theme.spacing(1)}px ${theme.spacing(2)}px`,
-+  padding: theme.spacing(1, 2), // '8px 16px'
+You can use the arguments to reduce the boilerplate. Instead of doing:
+
+```js
+padding: `${theme.spacing(1)}px ${theme.spacing(2)}px`, // '8px 16px'
+```
+
+you can do:
+
+```js
+padding: theme.spacing(1, 2), // '8px 16px'
 ```
