+++ docs/src/pages/customization/typography/typography.md
@@ -2,14 +2,11 @@
 
 <p class="description">The theme provides a set of type sizes that work well together, and also with the layout grid.</p>
 
-The following example demonstrates how to change the [default values](/customization/default-theme/?expend-path=$.typography) of the typography in the theme.
-You can learn more about how to use the [Typography component](/components/typography/) by checking out the dedicated page.
-
-{{"demo": "pages/customization/typography/TypographyTheme.js"}}
-
 ## Font family
 
-You can use the system font instead of the default Roboto font.
+You can change the font family with the `theme.typography.fontFamily` property.
+
+For instance, this demo use the system font instead of the default Roboto font:
 
 ```js
 const theme = createMuiTheme({
@@ -30,7 +27,7 @@ const theme = createMuiTheme({
 });
 ```
 
-## Self-host fonts
+### Self-host fonts
 
 To self-host fonts, download the font files in `ttf`, `woff`, and/or `woff2` formats and import them into your code.
 
@@ -55,25 +52,13 @@ const raleway = {
 };
 ```
 
-Then, you can change the theme to use this new font.
-In order to globally define Raleway as a font face the [`CssBaseline`](/components/css-baseline/) component needs to be used.
+Then, you need to change the theme to use this new font.
+In order to globally define Raleway as a font face the [`CssBaseline`](/components/css-baseline/) component can be used (or any other CSS solution of your choice).
 
 ```js
 const theme = createMuiTheme({
   typography: {
-    fontFamily: [
-      'Raleway',
-      '-apple-system',
-      'BlinkMacSystemFont',
-      '"Segoe UI"',
-      'Roboto',
-      '"Helvetica Neue"',
-      'Arial',
-      'sans-serif',
-      '"Apple Color Emoji"',
-      '"Segoe UI Emoji"',
-      '"Segoe UI Symbol"',
-    ].join(','),
+    fontFamily: 'Raleway, Arial',
   },
   overrides: {
     MuiCssBaseline: {
@@ -89,7 +74,7 @@ const theme = createMuiTheme({
 
 Material-UI uses `rem` units for the font size.
 The browser `<html>` element default font size is `16px`, but browsers have an option to change this value,
-so `rem` units allow us to accommodate the user's settings, resulting in a much better user experience.
+so `rem` units allow us to accommodate the user's settings, resulting in a better accessibility support.
 Users change font size settings for all kinds of reasons, from poor eyesight to choosing optimum settings
 for devices that can be vastly different in size and viewing distance.
 
@@ -111,7 +96,7 @@ The computed font size by the browser follows this mathematical equation:
 ![font-size](/static/images/font-size.gif)
 <!-- https://latex.codecogs.com/gif.latex?computed&space;=&space;specification&space;\frac{typography.fontSize}{14}&space;\frac{html&space;font&space;size}{typography.htmlFontSize} -->
 
-## HTML font size
+### HTML font size
 
 You might want to change the `<html>` element default font size. For instance, when using the [10px simplification](https://www.sitepoint.com/understanding-and-using-rem-units-in-css/).
 We provide a `htmlFontSize` theme property for this use case.
@@ -137,7 +122,7 @@ html {
 
 {{"demo": "pages/customization/typography/FontSizeTheme.js"}}
 
-## Responsive font sizes
+### Responsive font sizes
 
 The typography variants properties map directly to the generated CSS.
 You can use [media queries](/customization/breakpoints/#api) inside them:
@@ -162,8 +147,57 @@ To automate this setup, you can use the [`responsiveFontSizes()`](/customization
 
 You can see this in action in the example below. adjust your browser's window size, and notice how the font size changes as the width crosses the different [breakpoints](/customization/breakpoints/):
 
+```js
+import { createMuiTheme, responsiveFontSizes } from '@material-ui/core/styles';
+
+let theme = createMuiTheme();
+theme = responsiveFontSizes(theme);
+```
+
 {{"demo": "pages/customization/typography/ResponsiveFontSizes.js"}}
 
-## Fluid font sizes
+### Fluid font sizes
 
 To be done: [#15251](https://github.com/mui-org/material-ui/issues/15251).
+
+## Variants
+
+The typography object comes with [13 variants](/components/typography/#component) by default:
+
+- h1
+- h2
+- h3
+- h4
+- h5
+- h6
+- subtitle1
+- subtitle2
+- body1
+- body2
+- button
+- caption
+- overline
+
+Each of these variants can be customized individually:
+
+```js
+const theme = createMuiTheme({
+  typography: {
+    subtitle1: {
+      fontSize: 12,
+    },
+    body1: {
+      fontWeight: 500,
+    },
+    button: {
+      fontStyle: 'italic',
+    },
+  },
+});
+```
+
+{{"demo": "pages/customization/typography/TypographyVariants.js"}}
+
+## The default values
+
+You can explore all the default values of the typography using our [explorer](/customization/default-theme/?expend-path=$.typography) or by opening the dev tools console on this page (`window.theme.typography`).
