+++ packages/material-ui/src/styles/createTypography.js
@@ -28,6 +28,7 @@ export default function createTypography(palette, typography) {
     htmlFontSize = 16,
     // Apply the CSS properties to all the variants.
     allVariants,
+    pxToRem: pxToRem2,
     ...other
   } = typeof typography === 'function' ? typography(palette) : typography;
 
@@ -37,8 +38,8 @@ export default function createTypography(palette, typography) {
     `Material-UI: 'htmlFontSize' is required to be a number.`,
   );
 
-  const coef = fontSize / htmlFontSize;
-  const pxToRem = size => `${(size / htmlFontSize) * coef}rem`;
+  const coef = fontSize / 14;
+  const pxToRem = pxToRem2 || (size => `${(size / htmlFontSize) * coef}rem`);
   const buildVariant = (fontWeight, size, lineHeight, letterSpacing, casing) => ({
     fontFamily,
     fontWeight,
@@ -74,7 +75,7 @@ export default function createTypography(palette, typography) {
     {
       htmlFontSize,
       pxToRem,
-      round,
+      round, // TODO To remove in v5?
       fontFamily,
       fontSize,
       fontWeightLight,
